+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add support for linker property `ModuleDefinitionFile`.
+
 ## [1.4.0] - 2026-03-01
 
 ### Added